Implements Drawer Menu. resoleved #64.
