+++ package.json
@@ -11,6 +11,7 @@
     "upgrade-interactive": "npm-check --update"
   },
   "dependencies": {
+    "@mui/icons-material": "^5.14.3",
     "@mui/joy": "^5.0.0-beta.2",
     "@mui/material": "^5.14.5",
     "@mui/styled-engine-sc": "^5.12.0",
