+++ src/feature/Drawer/Index.ts
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+
+export const menuContext = createContext({
+  isOpened: false,
+  setOpened: (isOpen: boolean) => { }
+});
